requirements.txt
aiohttp==3.7.4.post0
async-timeout==3.0.1
asyncio==3.4.3
attrs==21.4.0
certifi==2021.10.8
cffi==1.15.0
chardet==4.0.0
charset-normalizer==2.0.10
dadjokes==1.3.2
DateTime==4.3
discord==1.7.3
discord.py==1.7.3
greenlet==1.1.2
idna==3.3
IMDbPY==2021.4.18
lxml==4.7.1
multidict==5.2.0
Pillow==9.0.0
pycparser==2.21
PyNaCl==1.5.0
pypng==0.0.21
PyQRCode==1.2.1
python-dotenv==0.19.2
pytz==2021.3
random-dog==1.0.2
requests==2.27.1
SQLAlchemy==1.4.29
typing_extensions==4.0.1
urllib3==1.26.8
yarl==1.7.2
youtube-dl==2021.12.17
zope.interface==5.4.0
